Time: 0 ms (100.00%) | Memory: 13.9 MB (58.14%) - LeetSync

+++ 260-single-number-iii/single-number-iii.cpp
@@ -7,7 +7,7 @@ public:
         {
             xor1=xor1^nums[i];
         }
-        long long right_bit = xor1 ^ (xor1 & (xor1-1));
+        int right_bit = int(xor1 ^ (xor1 & (xor1-1)));
         int b1=0;
         int b2=0;
         for(int i=0;i<n;i++)
